PyChat_v3-master/requirements.txt
ttkthemes
pycryptodome
